Update: Modularized Codebase and implemented aload_data method
