+++ llama-index-integrations/readers/llama-index-readers-gemini/llama_index/readers/gemini/__init__.py
@@ -1,11 +1,3 @@
-from llama_index.readers.gemini.base import (
-    GeminiReader,
-    TextChunk,
-    Chunks,
-    ProcessingStats,
-    CacheItem,
-    Document
-)
+from llama_index.readers.gemini.base import GeminiReader
 
-__all__ = ["GeminiReader", "TextChunk", "Chunks",
-           "ProcessingStats", "CacheItem", "Document"]
+__all__ = ["GeminiReader"]
